+++ strgicmp.h
@@ -4,6 +4,6 @@
 #include "autotrace.h"
 
 extern at_bool strgicmp (const char *s1, const char *s2);
-extern at_bool strgnicmp (const char *s1, const char *s2, long len);
+extern at_bool strgnicmp (const char *s1, const char *s2, size_t len);
 
 #endif /* not STRGICMP_H */
